requirements.txt
django
djangorestframework
requests
django-redis